requirements.txt
streamlit>=1.35.0
google-adk>=0.6.0
cairosvg>=2.7.0        # optional — for Python SVG→PNG conversion
